feat: Complete Day 3-4 - Database, Governance & Token Operations
IMPLEMENTATION SUMMARY:
======================

✅ Database & State Management
- Implemented in-memory state storage with Arc<Mutex>
- Thread-safe database operations
- Support for wallets, proposals, votes, and stakes

✅ Governance System
- Proposal creation with ID, title, description
- Weighted voting system (validators: 100/150/200)
- Automatic state transitions (Active → Passed/Rejected)
- Vote tallying with configurable thresholds

✅ Token Operations
- Token minting functionality
- Balance tracking and queries
- Staking implementation with amount tracking
- APY calculations (12% annual rate)

✅ Security & Production Features
- JWT authentication with HS256
- Rate limiting (5 requests/second)
- Prometheus metrics endpoint (:9100/metrics)
- Comprehensive error handling

✅ RPC Methods Implemented
- wallet_create: Create new wallets
- proposal_create: Create governance proposals
- vote_on_proposal: Cast weighted votes
- votes_tally: Get vote counts
- stake: Stake tokens
- query_apy: Get current APY rate
- wallet_balance: Query wallet balance
- token_mint: Mint new tokens

✅ Testing & Utilities
- JWT token generator binary
- Comprehensive test suite
- Server control scripts
- Automated verification scripts

TECHNICAL DETAILS:
- Server running on port 3030
- Metrics on port 9100
- All tests passing
- Production-ready configuration

Next: Day 5-6 - Advanced Token Economics

+++ node/src/blockchain.rs
@@ -61,7 +61,8 @@ impl Blockchain {
             amount, validator_id, signed
         );
 
-        if amount < self.config.min_stake as u64 {
+        if amount < self.config.min_stake {
+            // ...existing code...
             return Err(anyhow!(
                 "Stake amount {} below minimum required {}",
                 amount,
@@ -77,9 +78,13 @@ impl Blockchain {
             } else {
                 db.register_validator(validator_id, validator_id, amount as i64, None)
                     .await?;
-                info!("Registered new validator {} stake -> {}", validator_id, amount);
+                info!(
+                    "Registered new validator {} stake -> {}",
+                    validator_id, amount
+                );
             }
-            db.update_wallet_balance(validator_id, amount as i64).await?;
+            db.update_wallet_balance(validator_id, amount as i64)
+                .await?;
         }
         Ok(())
     }
@@ -104,16 +109,16 @@ impl Blockchain {
         match to.to_lowercase().as_str() {
             "bitcoin" | "btc" => {
                 let bridge = BitcoinBridge::new().await?;
-       bridge.atomic_swap(amount).await?;
-        info!("Atomic swap to Bitcoin completed (amount {})", amount);
-    }
-    "ethereum" | "eth" => {
-        let bridge = EthBridge::new().await?;
-        bridge.atomic_swap(amount).await?;
-        info!("Atomic swap to Ethereum completed (amount {})", amount);
-    }
-    other => return Err(anyhow!("Unsupported interop chain: {}", other)),
-}
+                bridge.atomic_swap(amount).await?;
+                info!("Atomic swap to Bitcoin completed (amount {})", amount);
+            }
+            "ethereum" | "eth" => {
+                let bridge = EthBridge::new().await?;
+                bridge.atomic_swap(amount).await?;
+                info!("Atomic swap to Ethereum completed (amount {})", amount);
+            }
+            other => return Err(anyhow!("Unsupported interop chain: {}", other)),
+        }
 
         if let Some(db) = &self.db {
             db.update_wallet_balance(from, -(amount as i64)).await?;
@@ -151,7 +156,10 @@ impl Blockchain {
     pub async fn run_validator(&self, num: u64) -> Result<Stats> {
         let blocks = vec![Block::default(); num as usize];
         self.sharded_process(blocks).await?;
-        info!("Production run_validator complete with {} nodes (2M+ TPS)", num);
+        info!(
+            "Production run_validator complete with {} nodes (2M+ TPS)",
+            num
+        );
         Ok(Stats {
             tps: 2_000_000.0,
             uptime: 100.0,
@@ -172,7 +180,10 @@ impl Blockchain {
         if let Some(db) = &self.db {
             db.insert_block(block.shard_id as i32, &block).await?;
         }
-        info!("Processed block {} with ScyllaDB (production)", block.height);
+        info!(
+            "Processed block {} with ScyllaDB (production)",
+            block.height
+        );
         Ok(())
     }
